% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_inscribed_r}
\alias{calc_inscribed_r}
\title{Calculate the radius of inscribed shapes}
\usage{
calc_inscribed_r(shape1, shape2, r1)
}
\arguments{
\item{shape1}{the 'outer' shape, character}

\item{shape2}{the 'inner' shape which we will calculate the radius for, character}

\item{r1}{the radius of the 'outer' shape, numeric}
}
\value{
the radius of the inner shape
}
\description{
Calculate the radius of inscribed shapes
}
