% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_inscribed}
\alias{choose_inscribed}
\title{Choose the number of inscribed planets}
\usage{
choose_inscribed(i, probs)
}
\arguments{
\item{probs}{}
}
\description{
Choose the number of inscribed planets
}
