% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glitch_spike}
\alias{glitch_spike}
\title{Title}
\usage{
glitch_spike(input, min_spikes = 30, max_spikes = 60,
  min_spike_r = 0.5, max_spike_r = 2, min_spike_jitter = -0.2,
  max_spike_jitter = 0.2)
}
\arguments{
\item{max_spike_jitter}{}
}
\value{

}
\description{
Title
}
