% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{inset}
\alias{inset}
\title{Add an inset plot}
\usage{
inset(spec, inset, position = "top-right")
}
\arguments{
\item{spec}{The main glyph_spec}

\item{inset}{A glyph_spec to place as an inset}

\item{position}{Where to place: "top-right", "top-left", "bottom-right",
"bottom-left", or a list(x, y, width, height) with proportions 0-1}
}
\value{
Modified glyph_spec
}
\description{
Add an inset plot
}
