% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makedf_inscribed_planets}
\alias{makedf_inscribed_planets}
\title{Calculate positions of inscribed planets}
\usage{
makedf_inscribed_planets(shape, r, size)
}
\arguments{
\item{shape}{the shape that the planets lie on, character}

\item{r}{the radius of the shape, numeric}

\item{size}{the size of the planets, numeric}
}
\value{
a dataframe with x and y positions and point size
}
\description{
Calculate positions of inscribed planets
}
