% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makedf_line}
\alias{makedf_line}
\title{Calculate x and y positions for a path}
\usage{
makedf_line(shape, r, line, width)
}
\arguments{
\item{shape}{the shape to draw, character}

\item{r}{the radius of the shape, numeric}

\item{line}{the type of line, character}

\item{width}{the width of the line, numeric}
}
\value{
a dataframe with x and y positions and associated parameters
}
\description{
Calculate x and y positions for a path
}
