% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makedf_orbits}
\alias{makedf_orbits}
\title{Make a dataframe of orbit positions}
\usage{
makedf_orbits(nlines, r, linetype, linewidth)
}
\arguments{
\item{nlines}{number of orbits, numeric}

\item{r}{radius of orbits, numeric}

\item{linetype}{type of line for orbits, character}

\item{linewidth}{width of the orbits, numeric}
}
\description{
Make a dataframe of orbit positions
}
