% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makedf_outlines}
\alias{makedf_outlines}
\title{Make a dataframe of outline positions}
\usage{
makedf_outlines(nlines, r, shapes, linetype, width)
}
\arguments{
\item{nlines}{the number of outlines, numeric}

\item{r}{the radius of the outlines, numeric}

\item{shapes}{the shape of the outlines, character}

\item{linetype}{the type of line for the outlines, character}

\item{width}{the width of the outlines, numeric}
}
\value{
a dataframe with x and y positions and associated parameters
}
\description{
Make a dataframe of outline positions
}
