% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makedf_planets}
\alias{makedf_planets}
\title{Make a dataframe for planet positions}
\usage{
makedf_planets(cen_x, cen_y, r)
}
\arguments{
\item{cen_x}{the center x coordinate of each planet, numeric}

\item{cen_y}{the center y coordinate of each planet, numeric}

\item{r}{the radius of each planet, numeric}
}
\description{
Make a dataframe for planet positions
}
