% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makedf_seed}
\alias{makedf_seed}
\title{Calculate positions for seed}
\usage{
makedf_seed(seed_shape, seed_r)
}
\arguments{
\item{seed_shape}{possible seed shapes, character}

\item{seed_r}{seed radius, numeric}
}
\value{
a data frame with x and y positions of the seed polygon
}
\description{
Calculate positions for seed
}
