% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{mark_area}
\alias{mark_area}
\title{Add an area mark}
\usage{
mark_area(spec, ..., data = NULL, style = list())
}
\arguments{
\item{spec}{A glyph_spec}

\item{...}{Aesthetic mappings (x, y, color, size, shape, alpha, tooltip)}

\item{data}{Optional per-mark data override}

\item{style}{Named list of fixed visual properties}
}
\value{
Modified \code{glyph_spec} object with the area mark added
}
\description{
Add an area mark
}
