% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{mark_bar}
\alias{mark_bar}
\title{Add a bar mark}
\usage{
mark_bar(spec, ..., data = NULL, style = list(), orient = "vertical")
}
\arguments{
\item{spec}{A glyph_spec}

\item{...}{Aesthetic mappings (x, y, color, size, shape, alpha, tooltip)}

\item{data}{Optional per-mark data override}

\item{style}{Named list of fixed visual properties}

\item{orient}{"vertical" or "horizontal"}
}
\value{
Modified \code{glyph_spec} object with the bar mark added
}
\description{
Add a bar mark
}
