% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{mark_point}
\alias{mark_point}
\title{Add a point mark (scatterplot)}
\usage{
mark_point(spec, ..., data = NULL, style = list())
}
\arguments{
\item{spec}{A glyph_spec}

\item{...}{Aesthetic mappings (x, y, color, size, shape, alpha, tooltip)}

\item{data}{Optional per-mark data override}

\item{style}{Named list of fixed visual properties}
}
\value{
Modified glyph_spec
}
\description{
Add a point mark (scatterplot)
}
\examples{
spec <- glyph(mtcars, x = wt, y = mpg) |>
  mark_point(color = cyl, size = hp, tooltip = "{cyl} cyl, {mpg} mpg")
}
