% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbit.R
\name{orbit}
\alias{orbit}
\title{Draw an orbital diagram}
\usage{
orbit(seed_probs = c(0.3, 0.5, 0.1, 0.1), planet_probs = c(0.4, 0.3,
  0.2, 0.2), pareto2_prob = c(0.2, 0.8))
}
\arguments{
\item{seed_probs}{probabilities of choosing each shape (none, circle, diamond, square), numeric vector of length 4}

\item{planet_probs}{probabilities of how many planet (0, 1, 2, 3), numeric vector of length 4}

\item{pareto2_prob}{probabilities of having a second set of pareto rings, numeric vector of length 2}
}
\value{
a ggplot of the orbital diagram
}
\description{
Draw an orbital diagram
}
