% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbit_glitch_app.R
\name{orbit_glitch_app}
\alias{orbit_glitch_app}
\title{Generate "glitched" orbital glyphs///FOR APP}
\usage{
orbit_glitch_app(seed_probs = c(1, 0, 0, 0), glitch_type = "spike",
  pareto2_prob = c(0.2, 0.8), glitch_params = list(num_glitches = 10,
  glitch_r_min = 0.5, glitch_r_max = 2, min_spikes = 30, max_spikes = 60,
  min_spikes2 = 15, max_spikes2 = 40, min_spike_jitter = -0.2,
  max_spike_jitter = 0.2))
}
\arguments{
\item{seed_probs}{probabilities of choosing each shape (none, circle,
diamond, square), numeric vector of length 4}

\item{glitch_type}{the type of glitch you want to introduce (options:
"spike", "connected", "shattered"), character}

\item{pareto2_prob}{probabilities of having a second set of pareto rings,
numeric vector of length 2}

\item{glitch_params}{a list of parameters for the glitch options, all
parameters must be specified, but only ones pertient to the glitch_type
will be used}
}
\value{
a glitched orbital ggplot
}
\description{
Generate "glitched" orbital glyphs///FOR APP
}
