% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{print.glyph_layout}
\alias{print.glyph_layout}
\title{Print a glyph layout}
\usage{
\method{print}{glyph_layout}(x, ...)
}
\arguments{
\item{x}{A glyph_layout object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the \code{glyph_layout} object
}
\description{
Print a glyph layout
}
