% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render a glyph spec as an htmlwidget}
\usage{
render(spec, width = NULL, height = NULL)
}
\arguments{
\item{spec}{A glyph_spec, glyph_layout, or glyph_compiled}

\item{width}{Widget width}

\item{height}{Widget height}
}
\value{
An \code{htmlwidget} object that renders the visualization
in an HTML viewer
}
\description{
Render a glyph spec as an htmlwidget
}
