% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scales}
\alias{scales}
\alias{scale_color}
\alias{scale_log}
\alias{scale_time}
\title{Scales}
\usage{
scale_color(spec, palette = "Tableau10", ...)

scale_log(spec, aesthetic = "y", base = 10, ...)

scale_time(spec, aesthetic = "x", ...)
}
\arguments{
\item{spec}{A glyph_spec}

\item{palette}{A named palette: "viridis", "Set2", "Tableau10", "Blues", etc.}

\item{...}{Additional arguments passed to \code{scale()}}

\item{aesthetic}{Which channel: "x", "y", "color", "size", "shape", "alpha"}

\item{base}{Log base (default 10)}
}
\value{
Modified \code{glyph_spec} object

Modified \code{glyph_spec} object

Modified \code{glyph_spec} object
}
\description{
Scales map data values to visual properties. Glyph simplifies
ggplot2's \verb{scale_<aesthetic>_<type>} explosion into a composable system:
\verb{scale(<aesthetic>, <type>, ...)}.
}
\section{Functions}{
\itemize{
\item \code{scale_color()}: Set color palette by name

\item \code{scale_log()}: Log-transform an axis

\item \code{scale_time()}: Time/date axis

}}
