% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph.R
\name{summary.glyph_spec}
\alias{summary.glyph_spec}
\title{Summarize a glyph specification}
\usage{
\method{summary}{glyph_spec}(object, ...)
}
\arguments{
\item{object}{A glyph_spec object}

\item{...}{Additional arguments (ignored)}
}
\value{
No return value, called for side effects (prints a summary
of the specification to the console)
}
\description{
Summarize a glyph specification
}
