% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{titles}
\alias{titles}
\title{Add a title, subtitle, or caption}
\usage{
titles(spec, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{spec}{A glyph_spec}

\item{title}{Main title}

\item{subtitle}{Subtitle (below title)}

\item{caption}{Caption (bottom of plot, for source attribution)}
}
\value{
Modified \code{glyph_spec} object with updated title metadata
}
\description{
Add a title, subtitle, or caption
}
