% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{to_vegalite}
\alias{to_vegalite}
\title{Export the spec as Vega-Lite JSON (interop)}
\usage{
to_vegalite(spec)
}
\arguments{
\item{spec}{A glyph_spec}
}
\value{
A JSON string in Vega-Lite format
}
\description{
Because glyph's spec is structurally similar to Vega-Lite,
we can export to Vega-Lite JSON for use in Python (Altair), JavaScript,
or the Vega Editor.
}
